% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conver_vcf.R
\name{convert_vcf}
\alias{convert_vcf}
\title{Convert vcf data into plink format}
\usage{
convert_vcf(vcf, outputfile = "subbackToPLINK")
}
\arguments{
\item{vcf}{A vcf.gz format to read into the function}

\item{outputfile}{file name when write out}
}
\value{
a plink format with .map .bim .fam .bed .SetID for Generate_SSD_SetID
}
\description{
read vcf.gz and transform into Plink to do SAKT analysis
}
\examples{
\dontrun{
dat<-system.file("extdata", "sub10k.vcf.gz", package = "skatvcf")
plink_data<-convert_vcf(vcf=dat, outputfile="subbackToPLINK")
}

}
