% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{functional_analysis}
\alias{functional_analysis}
\title{statistical analysis with SNP information in addition to meta data}
\usage{
functional_analysis(
  data,
  method = c("fisher", "chisq", "glm", "survfit", "coxph"),
  formula
)
}
\arguments{
\item{data}{a data with SNP and meta data}

\item{method}{statistical methods including "fisher", "chisq", "glm", "survfit", "coxph"}

\item{formula}{a symbolic descripton of the model to be fitted}
}
\value{
a table containing varaibles and p-value
}
\description{
users can choose different statistical methods to test significant variants
}
\examples{
dat<-read.csv(system.file("extdata", "merged_clinical_variant_data1_md.csv", package = "CLINSKAT"),
             stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
sur_model<-as.formula("survival::Surv(Follow_up_months, Last_Status) ~ X3.142172064C.G")
KM_result<-functional_analysis(data=dat, method="survfit", formula = sur_model)

dat$Gender<-as.factor(dat$Gender)
dat$X1.161298236A.G<-as.factor(dat$X1.161298236A.G)
mt<-table(dat$Gender, dat$X1.161298236A.G)
fisher_test<-functional_analysis(data=mt, method="fisher")

}
