% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatetoDisease.R
\name{relate2GeneDisease}
\alias{relate2GeneDisease}
\title{relate significant variants to pathways in MSigDB database}
\usage{
relate2GeneDisease(
  SNPdata,
  species = c("Homo sapiens", "Mus musculus"),
  category = c("H", "C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8")
)
}
\arguments{
\item{SNPdata}{significant variants of interest from statistical analysis}

\item{species}{model organisms such as hamun and mouse}

\item{category}{gene sets including "H", "C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8"}
}
\value{
a table including pathways, gene ratio, background ratio, and p-value
}
\description{
relate a list of genes, convert ensemble ID to gene ID and pathway enrichment analysis
}
\examples{
dat<-read.csv(system.file("extdata", "asso001.csv", package = "skatvcf"),
             stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
gene_disease<-relate2GeneDisease(SNPdata=dat, species="Homo sapiens", category="H")

}
