% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatetoDisease.R
\name{relatetoDisease}
\alias{relatetoDisease}
\title{relate a list of genes to disease table based on Clinvar}
\usage{
relatetoDisease(genelist)
}
\arguments{
\item{genelist}{genes of interest corresponding to disease table}
}
\value{
a table including genes and corresponding diseases
}
\description{
relate a list of genes to disease table based on Clinvar
}
\examples{
genes<-c("BCOR",  "NAA10")
gene_disease<-relatetoDisease(genelist=genes)

}
