% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skat_assoc.R
\name{skat_assoc}
\alias{skat_assoc}
\title{Convert vcf data into plink format}
\usage{
skat_assoc(
  vcf,
  outputfile = "subbackToPLINK",
  formula,
  out_type = "C",
  SNP_weight = NULL
)
}
\arguments{
\item{vcf}{A vcf.gz format to read into the function}

\item{outputfile}{file name when write out}

\item{formula}{an object of class formula}

\item{out_type}{an indicator of the outcome type. "C" for the continuous outcome and "D" for
the dichotomous outcome}

\item{SNP_weight}{output object from Read_SNP_WeightFile (default=NULL)}
}
\value{
a plink format with .map .bim .fam .bed .SetID for Generate_SSD_SetID
}
\description{
read vcf.gz and transform into Plink to do SKAT analysis
}
\examples{
\dontrun{
dat<-system.file("extdata", "sub10k.vcf.gz", package = "CLINSKAT")
formula=as.formula(y~1)
plink_data<-skat_assoc(vcf=dat, outputfile="subbackToPLINK", formula=formula, out_type="C")
}

}
