% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_test.R
\name{statistical_test}
\alias{statistical_test}
\title{statistical analysis with SNP information in addition to meta data}
\usage{
statistical_test(
  data,
  method = c("fisher", "chisq", "lm", "survfit", "coxph"),
  formula
)
}
\arguments{
\item{data}{a data with SNP and meta data}

\item{method}{statistical methods including "fisher", "chisq", "lm", "survfit", "coxph"}

\item{formula}{a symbolic descripton of the model to be fitted}
}
\value{
a table containing varaibles and p-value
}
\description{
users can choose different statistical methods to test significant variants
}
\examples{
dat<-read.csv(system.file("extdata", "merged_clinical_variant_data1_md.csv", package = "skatvcf"),
             stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
sur_model<-as.formula("survival::Surv(Follow_up_months, Last_Status) ~ X3.142172064C.G")
KM_result<-statistical_test(data=dat, method="survfit", formula = sur_model)

dat$Gender<-as.factor(dat$Gender)
dat$X1.161298236A.G<-as.factor(dat$X1.161298236A.G)
mt<-table(dat$Gender, dat$X1.161298236A.G)
fisher_test<-statistical_test(data=mt, method="fisher")

}
